Modify delete tool: check if request comes from the owner of the tool

+++ src/main/java/com/toolborrow/backend/service/ToolServiceImpl.java
@@ -17,14 +17,17 @@ import com.toolborrow.backend.repository.UserRepository;
 import com.toolborrow.backend.utils.JwtUtils;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
+import static org.springframework.http.HttpStatus.*;
 
 @Service
 @RequiredArgsConstructor
@@ -126,7 +129,16 @@ public class ToolServiceImpl implements ToolService {
             throw new TBAException(BAD_REQUEST, "Tool has active reservations and cannot be deleted: " + id);
         }
 
-        toolRepository.deleteById(id);
+        Optional<Tool> tool = toolRepository.findById(id);
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = (String) authentication.getPrincipal();
+
+        if (!Objects.equals(email, tool.get().getUser().getEmail())) {
+            throw new TBAException(BAD_REQUEST, "Only the owner of the tool can remove it!");
+        } else {
+            toolRepository.deleteById(id);
+        }
     }
 
     @Override
